Change to have separate classes for resolve to, and resolve expectations
